requirements.txt
# Qlib 核心依赖（基于官方 examples/benchmarks 的 requirements.txt）
pyqlib
numpy>=1.21.0
pandas>=1.1.2
scikit-learn>=0.23.2

# 数据采集模块依赖（从 alphaHome 迁移）
asyncpg>=0.27.0        # PostgreSQL 异步数据库驱动
psycopg2-binary>=2.9.0 # PostgreSQL 同步数据库驱动
aiohttp>=3.8.0         # 异步 HTTP 客户端（Tushare API）
tushare>=1.2.89        # Tushare 数据源
appdirs>=1.4.4         # 应用程序目录工具（配置管理）
aiolimiter>=1.0.0      # 异步速率限制
python-dotenv>=1.0.0   # 环境变量管理（.env文件支持）

# GUI 模块依赖
async-tkinter-loop>=0.9.0  # Tkinter 异步事件循环支持
tkcalendar>=1.6.0          # 日期选择器组件（可选）

# 可选：深度学习模型
# torch==1.7.0

# 可选：Jupyter 支持
# jupyter
# notebook

# 可选：可视化
# matplotlib
# plotly
